Cleanup: remove unnecessary boxing/unboxing

+++ mdsalutil/mdsalutil-api/src/main/java/org/opendaylight/genius/mdsalutil/InstructionType.java
@@ -53,7 +53,7 @@ public enum InstructionType {
             return new InstructionBuilder()
                     .setInstruction(
                             new GoToTableCaseBuilder().setGoToTable(
-                                    new GoToTableBuilder().setTableId(Short.valueOf(tableId)).build()).build())
+                                    new GoToTableBuilder().setTableId(tableId).build()).build())
                     .setKey(new InstructionKey(instructionKey)).build();
         }
     },
